Break up isoCrypto
