+++ packages/server/src/helpers/iso/isoCrypto/structs.ts
@@ -1,0 +1,2 @@
+export type SubtleCryptoAlg = "SHA-1" | "SHA-256" | "SHA-384" | "SHA-512";
+export type SubtleCryptoCrv = "P-256" | "P-384" | "P-521";
